Changes due to implementation of DBLookupTable
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@91 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/database/CopySQLData.java
@@ -172,7 +172,7 @@ public abstract class CopySQLData {
 		ListIterator iterator = fieldTypes.listIterator();
 
 		while (iterator.hasNext()) {
-			switch (((Short) iterator.next()).shortValue()) {
+			switch (((Integer) iterator.next()).shortValue()) {
 							case Types.CHAR:
 							case Types.LONGVARCHAR:
 							case Types.VARCHAR:
